Ajustado problema de clicar em alguma Meta; Alterado a Barra de Navegação; Alterado nome dos arquivos.
