+++ src/App.css
@@ -4,7 +4,6 @@
 
 .App-logo {
   height: 50vmin;
-
 }
 
 .App-header {
