+++ src/App.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import BarraNavegacao from './pages/BarraNavegacao/BarraNavegacao'
+import Routes from './routes'
 
 function App() {
   return (
       <div className="App">
-          <BarraNavegacao />
+          <Routes />
       </div>
   )
 }
