--- src/pages/Metas/Metas.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import MetasCurto from './MetasCurto'
-import MetasLongo from './MetasLongo'
-import MetasMedio from './MetasMedio'
-
-function Metas() {
-    return (
-        <div>
-            <MetasCurto />
-            <MetasMedio />
-            <MetasLongo />
-        </div>
-    )
-}
-export default Metas
