+++ src/pages/Metas/metaSelecionada.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import Apis from '../Api'
-import GridMeta from './Grid/GridMeta'
+import Apis from '../../services/api'
+import {Link} from 'react-router-dom'
 
-class Metas extends React.Component {
+export default class MetaSelecionada extends React.Component {
     constructor(props) {
         super(props)
 
@@ -20,37 +20,34 @@ class Metas extends React.Component {
 
     loadData() {
         this.setState({isLoading: true})
-        Apis.carregarMetas(this.state.metas).then((res)=>{
+        var user_id = localStorage.getItem('user')
+        Apis.carregarMetasPorId(user_id).then((res)=>{
             this.setState({
                 isLoading: false,
                 metas: res.data
             })
-        })
-    }
-
-    deletarMetas(id) {
-        Apis.deletarMetas(id).then((res)=> this.loadData())
+        }) 
     }
 
     renderMetas(metas){
         return (
-                <GridMeta   textPrimeiro={metas.titulo}
-                            textSegundo={metas.tipos}
-                            onClick={() => this.deletarMetas(metas.id)}
-                            key={metas.id}
-                />
+            <div key={metas.id}  >
+                <p>{metas.titulo}</p>
+                <p>{metas.tipos}</p>
+                <p>{metas.descricao}</p>
+            </div>
         )
     }
 
     render(metas) {
         return (
             <div>
-                <h1>Bem vindo as suas Metas</h1>
+                <h1>Metas de médio prazo</h1>
                 <div id="series" className="row list-group" >
                     {this.state.metas.map(this.renderMetas)}
-                </div>                    
+                </div>
+                <Link to='/metas'>Voltar</Link>                
             </div>
         )
     }
 }
-export default Metas
