+++ src/pages/Metas/metas.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import MetasCurto from './metasCurto'
+import MetasLongo from './metasLongo'
+import MetasMedio from './metasMedio'
+import BarraNavegacao from '../BarraNavegacao/barraNavegacao'
+
+export default function Metas() {
+    return (
+        <div>
+            <BarraNavegacao />
+            <MetasCurto />
+            <MetasMedio />
+            <MetasLongo />
+        </div>
+    )
+}
