+++ src/pages/Metas/metasMedio.js
@@ -1,8 +1,10 @@
 import React from 'react'
-import Apis from '../../services/Api'
-import GridMeta from '../../components/Grid/GridMeta'
+import Apis from '../../services/api'
+import GridMeta from '../../components/Grid/gridMeta'
+import { Link } from 'react-router-dom'
+import ButtonIcon from '../../components/Button/buttonIcon'
 
-class MetasMedio extends React.Component {
+export default class MetasMedio extends React.Component {
     constructor(props) {
         super(props)
 
@@ -28,17 +30,27 @@ class MetasMedio extends React.Component {
         }) 
     }
 
+    navegarMetaSelecionada(id) {
+        localStorage.setItem('user', id)
+    }
+
     deletarMetas(id) {
         Apis.deletarMetas(id).then((res)=> this.loadData())
     }
 
     renderMetas(metas){
         return (
-            <GridMeta   textPrimeiro={metas.titulo}
-                        textSegundo={metas.tipos}
-                        onClick={() => this.deletarMetas(metas.id)}
-                        key={metas.id}
-            />
+            <>
+                <Link to={'/meta_selecionada/' + metas.id} onClick={() => this.navegarMetaSelecionada(metas.id)}>
+                    <GridMeta   textPrimeiro={metas.titulo}
+                                textSegundo={metas.tipos}
+                                onClick={() => this.deletarMetas(metas.id)}
+                                key={metas.id}
+                                to={'/meta_selecionada/'}
+                    />
+                </Link>
+                <ButtonIcon onClick={() => this.deletarMetas(metas.id)}/>
+            </>
         )
     }
 
@@ -53,4 +65,3 @@ class MetasMedio extends React.Component {
         )
     }
 }
-export default MetasMedio
