+++ src/services/api.js
@@ -7,6 +7,7 @@ const Api = axios.create({
 export const carregarMetas = () => Api.get('metas')
 export const carregarTipos = () => Api.get('tipos')
 export const carregarMetasPorTipos = (tipos) => Api.get('metas?tipos='+tipos)
+export const carregarMetasPorId = (id) => Api.get('metas?id=' + id)
 export const salvarMetas = (novaMeta) => Api.post('metas', novaMeta)
 export const deletarMetas = (id) => Api.delete('metas/' + id)
 
@@ -15,6 +16,7 @@ export const Apis = {
      carregarTipos,
      salvarMetas,
      carregarMetasPorTipos,
+     carregarMetasPorId,
      deletarMetas
  }
